fix: rename footer plugin to match cammelCase
