+++ packages/wxt/src/core/builders/vite/plugins/iifeFooter.ts
