+++ packages/wxt/src/core/builders/vite/plugins/index.ts
@@ -13,4 +13,4 @@ export * from './defineImportMeta';
 export * from './removeEntrypointMainFunction';
 export * from './wxtPluginLoader';
 export * from './resolveAppConfig';
-export * from './iife-footer';
+export * from './iifeFooter';
